+++ EntityTest/CSphere.cpp
@@ -2,17 +2,17 @@
 
 #include <glm/gtx/norm.hpp>
 
-CSphere::CSphere( const float radius ) :
+CSphere::CSphere( const f16 radius ) :
 	m_radius { radius },
 	m_radius2 { radius * radius }
 {}
 
-const float &CSphere::Radius() const
+const f16 &CSphere::Radius() const
 {
 	return( m_radius );
 }
 
-const float &CSphere::Radius2() const
+const f16 &CSphere::Radius2() const
 {
 	return( m_radius2 );
 }
